requirements.txt
agentql==1.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
fastapi==0.115.2
filelock==3.16.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
identify==2.6.1
idna==3.8
iniconfig==2.0.0
instructor==1.3.7
Jinja2==3.1.4
jiter==0.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
multidict==6.1.0
nodeenv==1.9.1
numpy==2.1.2
openai==1.33.0
packaging==24.1
platformdirs==4.3.6
playwright==1.46.0
pluggy==1.5.0
pre_commit==4.0.1
propcache==0.2.0
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
pyee==11.1.0
Pygments==2.18.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-multipart==0.0.12
pyxel==2.2.1
PyYAML==6.0.2
requests==2.32.3
rich==13.8.1
setuptools==74.1.2
shellingham==1.5.4
sniffio==1.3.1
starlette==0.39.2
swarm @ git+https://github.com/openai/swarm.git@4b890df9e2f6daa02b8c9228ebc8f55abb6756a5
tenacity==8.5.0
tf-playwright-stealth==0.0.4
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.31.1
virtualenv==20.26.6
yarl==1.15.0
